+++ packages/sdk/src/models/record_store.ts
@@ -469,4 +469,5 @@ class RecordStore extends AbstractModelWithAsyncData<TableData, WatchableRecordS
     }
 }
 
+/** @internal */
 export default RecordStore;