+++ packages/sdk/src/models/view_data_store.ts
@@ -260,4 +260,5 @@ class ViewDataStore extends AbstractModelWithAsyncData<ViewData, WatchableViewDa
     }
 }
 
+/** @internal */
 export default ViewDataStore;